backend/requirements.txt
asgiref==3.3.1
Django==3.1.4
django-cors-headers==3.6.0
django-crispy-forms==1.10.0
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
mysqlclient==2.0.2
Pillow==8.0.1
PyJWT==2.0.0
python-decouple==3.4
pytz==2020.5
sqlparse==0.4.1
